+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kCDspBase/AbstractDmc4kCDspBaseAdapterSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCDspBase
+{
+	public abstract class AbstractDmc4kCDspBaseAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}