+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kCDspBase/AbstractDmc4kCDspBaseAdapterRoutingControl.cs
@@ -2,7 +2,7 @@
 namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCDspBase
 {
 	public abstract class AbstractDmc4kCDspBaseAdapterRoutingControl<TParent> : AbstractInputCardAdapterRoutingControl<TParent>
-		where TParent : ICardAdapter
+		where TParent : IInputCardAdapter
 	{
 		/// <summary>
 		/// Constructor.