+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdBase/AbstractDmc4kHdBaseAdapterRoutingControl.cs
@@ -2,7 +2,7 @@
 namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdBase
 {
 	public abstract class AbstractDmc4kHdBaseAdapterRoutingControl<TParent> : AbstractInputCardAdapterRoutingControl<TParent>
-		where TParent : ICardAdapter
+		where TParent : IInputCardAdapter
 	{
 		/// <summary>
 		/// Constructor.