+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdDspBase/AbstractDmc4kHdDspBaseAdapterRoutingControl.cs
@@ -2,7 +2,7 @@
 namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdDspBase
 {
 	public abstract class AbstractDmc4kHdDspBaseAdapterRoutingControl<TParent> : AbstractInputCardAdapterRoutingControl<TParent>
-		where TParent : ICardAdapter
+		where TParent : IInputCardAdapter
 	{
 		/// <summary>
 		/// Constructor.