+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+ - Added DM-TX-4KZ-202-C and DM-TX-4KZ-302-C transmitters
  - Added DMC-4K-HD and DMC-4KZ-HD input cards
 
 ## [14.1.0] - 2019-01-29