+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added DMC-4KZ-C-DSP input card
  - Added DMC-4KZ-C input card
  - Added DMC-4KZ-HD-DSP input card
+ - Added DMC-4KZ-CO-HD output card
 
 ## [14.1.0] - 2019-01-29
 ### Added